Natural order imposed and improved visualization.
Now nodes in the graph are added following the natural traversal order of a quantum circuit: the circuit is divided into layers (a layer contains only gates that act on separate qubits), and gates in each layer are ordered by qubit index. As a consequence, node_ids attribute contains identifiers of the nodes in this order, as well as node_feature_matrix should preserve this order: still node_mapping is available to ensure consistency between nodes and their features in the matrix.

draw() method added to the class to visualize the graph correspondent to the quantum circuit.

Class has been refactored a bit, for better organization and readability.