Refactor QuantumCircuitGraph to include adjacency matrix.
nx.adjacency_matrix cannot be used, as well as QuantumCircuitGraph.graph.nodes, sice are ordered by the order in which nodes are added to the graph: indeed there are some inconsistencies between addition order and the desired one, since control and target of a CNOT gate are added at the same time to the circuit. Then build_adjacency_matrix() method will produce the adjacency in the desired order, in CSR format.