Implemented CNOTDecompositionTransformation.
This transformation allows to decompose a single CNOT gate into a sequence of 4 CNOTs, using an ancilla qubit, and viceversa. The application of both transformations, forward and backward, is handled by identifying the particular pattern using the class attribute matching_key: when one of the available pattern is matched, this matching_key is updated so that the next transformation steps can use this information.