Implemented ChangeOfBasisTransformation.
This transformation implement the equivalence HXH=Z, which allows to perform a change of basis from the |0>, |1> basis to the |+>, |-> one, and all the deriving equivalences. To handle all these patterns, an attribute transformation_rules stores in a dictionary the mapping between possible patterns and the equivalent sequence of gates to be substituted.