Refactored method process_gate().
Now method process_gate() takes only care of extracting node label and information to be fed to insert_node() method: this last method, given a graph and the necessary information about the node to be added, insert the node in the graph. Now predecessor and successor in the graph of the current node can be provided, so that a gate can be inserted in the circuit in ay position, and not at the end only.