Refactored graph creation and visualization processes.
Now the graph representation of the circuit is always constructed in the order given by gate addition in the original circuit: this means no more addition to the graph by layer, that do not allow to define a mapping between node label and the corresponding gate index in the circuit's data attribute. Then beacomes necessary to refactor node positions computation, by creating a visualization_graph, constructed layer by layer while computing the positions of the nodes: when visualizing the graph representation in a 'circuit_like' format, the visualization_graph will be displayed (completely analogous to the graph stored in the already existing graph attribute).